Optimisation attribution point/pixel, erreur attribution PGM  à corriger

+++ point.cpp
@@ -1,9 +1,9 @@
 #include "point.h"
 #include <iostream>
-Point::Point(float lat, float longi, float alti): m_lat(lat),m_longi(longi), m_alti(alti)
+Point::Point(double lat, double longi, double alti): m_lat(lat),m_longi(longi), m_alti(alti)
 {   
-    m_x = 0.;
-    m_y = 0.;
+    m_x = 0;
+    m_y = 0;
     m_z = 0.;
 }
 void Point::projection(PJ *P, PJ_COORD coord_wgs84, PJ_COORD coord_lambert93)
@@ -18,23 +18,23 @@ void Point::projection(PJ *P, PJ_COORD coord_wgs84, PJ_COORD coord_lambert93)
     m_z = m_alti;
 }
 
-float Point::read_x()
+int Point::read_x()
 {
   return m_x;
 }
 
- float Point::read_y()
+ int Point::read_y()
 {
   return m_y;
 }
-float Point::read_z()
+double Point::read_z()
 {
   return m_z;
 }
 istream& operator>>(istream& stream, Point& pt)
 {
 
-  stream >>pt.m_lat; // recopiage du float
+  stream >>pt.m_lat; // recopiage du double
   stream >>pt.m_longi;
   stream >>pt.m_alti;
